requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.1


dj-rest-auth
Django
django-allauth
django-braces
django-classy-tags
django-configuration
django-configurations
django-cors-headers
django-countries
django-crispy-forms
django-custom-user
django-debug-toolbar
django-environ
django-extensions
django-filter
django-js-asset
django-mptt
django-nyt
django-polymorphic
django-sekizai
django-taggit
djangorestframework
djangorestframework-simplejwt
docstring-to-markdown 
docutils 
drf-nested-routers
drf-spectacular
drf-yasg
# transformers

typing_extensions
resend
gunicorn
Werkzeug
pycite
whitenoise
psycopg2-binary
colorama
Pillow
pytz
six
# transformers
